+++ X10D/src/Text/CharExtensions.cs
@@ -47,4 +47,34 @@ public static class CharExtensions
             _ => new string(value, count)
         };
     }
+
+    /// <summary>
+    ///     Writes a character to a span of characters, repeated a specified number of times.
+    /// </summary>
+    /// <param name="value">The character to repeat.</param>
+    /// <param name="count">The number of times to repeat.</param>
+    /// <param name="destination">The span of characters into which the repeated characters will be written.</param>
+    [MethodImpl(CompilerResources.MethodImplOptions)]
+    public static void Repeat(this char value, int count, Span<char> destination)
+    {
+        if (count < 0)
+        {
+            throw new ArgumentOutOfRangeException(nameof(count), ExceptionMessages.CountMustBeGreaterThanOrEqualTo0);
+        }
+
+        if (count == 0)
+        {
+            return;
+        }
+
+        if (destination.Length < count)
+        {
+            throw new ArgumentException(ExceptionMessages.DestinationSpanLengthTooShort, nameof(destination));
+        }
+
+        for (var i = 0; i < count; i++)
+        {
+            destination[i] = value;
+        }
+    }
 }