+++ X10D/src/Text/StringExtensions.cs
@@ -932,14 +932,50 @@ public static class StringExtensions
                 return value;
         }
 
-        var builder = new StringBuilder(value.Length * count);
+        Span<char> destination = stackalloc char[value.Length * count];
+        value.Repeat(count, destination);
+        return new string(destination);
+    }
 
-        for (var i = 0; i < count; i++)
+    /// <summary>
+    ///     Repeats a string a specified number of times, writing the result to a span of characters.
+    /// </summary>
+    /// <param name="value">The string to repeat.</param>
+    /// <param name="count">The repeat count.</param>
+    /// <param name="destination">The destination span to write to.</param>
+    /// <exception cref="ArgumentNullException"><paramref name="value" /> is <see langword="null" />.</exception>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="count" /> is less than 0.</exception>
+    /// <exception cref="ArgumentException">
+    ///     <paramref name="destination" /> is too short to contain the repeated string.
+    /// </exception>
+    [MethodImpl(CompilerResources.MethodImplOptions)]
+    public static void Repeat(this string value, int count, Span<char> destination)
+    {
+        if (value is null)
         {
-            builder.Append(value);
+            throw new ArgumentNullException(nameof(value));
         }
 
-        return builder.ToString();
+        if (count < 0)
+        {
+            throw new ArgumentOutOfRangeException(nameof(count), ExceptionMessages.CountMustBeGreaterThanOrEqualTo0);
+        }
+
+        if (count == 0)
+        {
+            return;
+        }
+
+        if (destination.Length < value.Length * count)
+        {
+            throw new ArgumentException(ExceptionMessages.DestinationSpanLengthTooShort, nameof(destination));
+        }
+
+        for (var iteration = 0; iteration < count; iteration++)
+        {
+            Span<char> slice = destination.Slice(iteration * value.Length, value.Length);
+            value.AsSpan().CopyTo(slice);
+        }
     }
 
     /// <summary>